chore: auth, logout repository 병합
